+++ e2e/tests/_helpers.js
@@ -1,5 +1,3 @@
-const assert = require('chai').assert
-
 exports.openMockFile = (server, mockFileName, fileContent) => {
   server.send({
     command: 'open',