+++ e2e/tests/completionEntryDetails.js
@@ -1,4 +1,3 @@
-const assert = require('chai').assert
 const path = require('path')
 const createServer = require('../server-fixture')
 const { openMockFile, getFirstResponseOfType } = require('./_helpers')