+++ e2e/tests/outliningSpans.js
@@ -1,6 +1,5 @@
 //@ts-check
 const path = require('path')
-const assert = require('chai').assert
 const createServer = require('../server-fixture')
 const { openMockFile, getFirstResponseOfType } = require('./_helpers')